refactor: refactor the builder into multiple files
